Biobjective Binary Programming Instances

Biobjective Binary Programs

Biobjective Assignment Problems (BOAP)

Modolib.read_boap_hadi_instance — Function.

read_boap_hadi_instance(num::Int64=1)

Returns a BOBPInstance corresponding to the numth Bi Objective Assignment Problem ( Hadi Instance )

Modolib.read_boap_hadi_non_dom_pts — Function.

read_boap_hadi_non_dom_pts(num::Int64=1)

Returns the Efficient Frontier of the numth Bi Objective Assignment Problem ( Hadi Instance )

Modolib.read_boap_hadi — Function.

read_boap_hadi(num::Int64)

Returns a BOBPInstance corresponding to the numth Bi Objective Assignment Problem ( Hadi Instance ) and its Efficient Frontier

Biobjective Knapsack Problems (BOKP)

Xavier Instances

read_bokp_xavier_instance1
read_bokp_xavier_instance2

Modolib.read_bokp_xavier_non_dom_pts — Function.

read_bokp_xavier_non_dom_pts(filename::AbstractString)

Returns Efficient Frontier of a Bi Objective Knapsack Problem ( Xavier Instance )

Modolib.read_bokp_xavier1 — Function.

read_bokp_xavier1(name::String)

Returns a BOBPInstance of a Bi Objective Knapsack Problem ( Xavier Class 1 Instance ) and its Efficient Frontier

Modolib.read_bokp_xavier2 — Function.

read_bokp_xavier2(name::String)

Returns a BOBPInstance of a Bi Objective Knapsack Problem ( Xavier Class 2 Instance ) and its Efficient Frontier

Hadi Instances

Modolib.read_bokp_hadi_instance — Function.

read_bokp_hadi_instance(num::Int64=1)

Returns a BOBPInstance corresponding to the numth Bi Objective Knapsack Problem ( Hadi Instance )

Modolib.read_bokp_hadi_non_dom_pts — Function.

read_bokp_hadi_non_dom_pts(num::Int64=1)

Returns the Efficient Frontier of the numth Bi Objective Knapsack Problem ( Hadi Instance )

Modolib.read_bokp_hadi — Function.

read_bokp_hadi(num::Int64)

Returns a BOBPInstance corresponding to the numth Bi Objective Knapsack Problem ( Hadi Instance ) and its Efficient Frontier

Biobjective Multi Dimensional Knapsack Problems (BOKP)

For this set of instances, the efficient frontier is not available.

Modolib.read_bokp_aritra_instance — Function.

read_bokp_aritra_instance(n::Int64, m::Int64, instance::Int64)

Returns a BOBPInstance of a Bi Objective Knapsack Problem ( Aritra Instances ) corresponding to n, m and instance.

Biobjective Set Covering Problems (BOSCP)

Modolib.read_boscp_xavier_instance — Function.

read_boscp_xavier_instance(instance_id::String)

Returns a BOBPInstance of a Bi Objective Set Covering Problem ( Xavier Instance ) corresponding to instance_id.

Modolib.read_boscp_xavier_non_dom_pts — Function.

read_boscp_xavier_non_dom_pts(instance_id::String)

Returns the Efficient Frontier of the Bi Objective Set Covering Problem ( Xavier Instance ) corresponding to instance_id.

Modolib.read_boscp_xavier — Function.

read_boscp_xavier(instance_id::String)

Returns a BOBPInstance of a Bi Objective Set Covering Problem ( Xavier Instance ) and its Efficient Frontier.

Biobjective Set Packing Problems (BOSPP)

Modolib.read_bospp_xavier_instance — Function.

read_bospp_xavier_instance(instance_id::String)

Returns a BOBPInstance of a Bi Objective Set Packing Problem ( Xavier Instance ) corresponding to filename.

Modolib.read_bospp_xavier_non_dom_pts — Function.

read_bospp_xavier_non_dom_pts(instance_id::String)

Returns the Efficient Frontier of a Bi Objective Set Packing Problem ( Xavier Instance )

Modolib.read_bospp_xavier — Function.

read_bospp_xavier(instance_id::String)

Returns a BOBPInstance of a Bi Objective Set Packing Problem ( Xavier Instance ) and its Efficient Frontier.