Biobjective Mixed Binary Programming Instances

Biobjective Mixed Binary Programs

Biobjective Mixed Binary Linear Programming Instance (BOMIP)

Modolib.read_bomip_hadi_instance — Function.

read_bomip_hadi_instance(num::Int64, ϵ::Float64=1.0e-9)

Returns a BOMBLPInstance corresponding to the numth Bi Objective Mixed Integer Programming Problem ( Hadi Instance )

Modolib.read_bomip_hadi_non_dom_pts — Function.

read_bomip_hadi_non_dom_pts(num::Int64)

Returns the Efficient Frontier of the numth Bi Objective Mixed Integer Programming Problem ( Hadi Instance )

Modolib.read_bomip_hadi — Function.

read_bomip_hadi(num::Int64, ϵ::Float64=1.0e-9)

Returns a BOMBLPInstance corresponding to the numth Bi Objective Mixed Integer Programming Problem ( Hadi Instance ) and its Efficient Frontier

Biobjective Uncapacitated Facility Location Problem (BOUFLP)

Modolib.read_bouflp_hadi_instance — Function.

read_bouflp_hadi_instance(num::Int64, ϵ::Float64=1.0e-9)

Returns a BOMBLPInstance corresponding to the numth Bi Objective Uncapacitated Facility Location Problem ( Hadi Instance )

Modolib.read_bouflp_hadi_non_dom_pts — Function.

read_bouflp_hadi_non_dom_pts(num::Int64)

Returns the Efficient Frontier of the numth Bi Objective Uncapacitated Facility Location Problem ( Hadi Instance )

Modolib.read_bouflp_hadi — Function.

read_bouflp_hadi(num::Int64, ϵ::Float64=1.0e-9)

Returns a BOMBLPInstance corresponding to the numth Bi Objective Uncapacitated Facility Location Problem ( Hadi Instance ) and its Efficient Frontier

Biobjective Best Subset Selection (BOBSS)

Modolib.generate_bobss_bomblpinstance — Function.

generate_bobss_bomblpinstance(y::Vector{Float64}, x::Array{Float64, 2}, M::Float64, ϵ::Float64=1.0e-9)

Returns a BOMBLPInstance for the Best Subset Selection Problem corresponding to y, x and M

Modolib.read_bobss_hadi_xycor — Function.

read_bobss_hadi_xycor(p::Int64, n::Int64, instance::Int64)

Returns y::Vector{Float64}, x::Array{Float64, 2} and correlation::Vector{Float64} corresponding to p, n and instance Bi Objective Best Subset Selection Problem ( Hadi Instance )

Modolib.read_bobss_hadi_non_dom_pts — Function.

read_bobss_hadi_non_dom_pts(p::Int64, n::Int64, instance::Int64)

Returns the Efficient Frontier of the Bi Objective Best Subset Selection Problem ( Hadi Instance ) corresponding to p, n and instance

Modolib.read_bobss_hadi — Function.

read_bobss_hadi(p::Int64, n::Int64, instance::Int64, M::Float64=1.0e10, ϵ::Float64=1.0e-9)

Returns y, x, correlation, BOMBLPInstance and the Efficient Frontier for the Best Subset Selection Problem ( Hadi Instance ) corresponding to y, x and M