Multiobjective Binary Programming Instances

Multiobjective Binary Programs

Multiobjective Assignment Problems (MOAP)

Modolib.read_moap_kirlik_instance — Function.

read_moap_kirlik_instance(p::Int64, n::Int64, instance::Int64)

Returns a MOBPInstance of a Multi Objective Assignment Problem ( Kirlik Instance ) corresponding to p, n and instance

Modolib.read_moap_kirlik_non_dom_pts — Function.

read_moap_kirlik_non_dom_pts(p::Int64, n::Int64, instance::Int64)

Returns the Efficient Frontier of the Multi Objective Assignment Problem ( Kirlik Instance ) corresponding to p, n and instance

Modolib.read_moap_kirlik — Function.

read_moap_kirlik(p::Int64, n::Int64, instance::Int64)

Returns MOBPInstance and the Efficient Frontier of the Multi Objective Assignment Problem ( Kirlik Instance ) corresponding to p, n and instance

Multiobjective Knapsack Problems (MOKP)

Modolib.read_mokp_kirlik_instance — Function.

read_mokp_kirlik_instance(p::Int64, n::Int64, instance::Int64)

Returns a MOBPInstance of a Multi Objective Knapsack Problem ( Kirlik Instance ) corresponding to p, n and instance

Modolib.read_mokp_kirlik_non_dom_pts — Function.

read_mokp_kirlik_non_dom_pts(p::Int64, n::Int64, instance::Int64)

Returns the Efficient Frontier of the Multi Objective Knapsack Problem ( Kirlik Instance ) corresponding to p, n and instance

Modolib.read_mokp_kirlik — Function.

read_mokp_kirlik(p::Int64, n::Int64, instance::Int64)

Returns MOBPInstance and the Efficient Frontier of the Multi Objective Knapsack Problem ( Kirlik Instance ) corresponding to p, n and instance